refactor ui trace canvas
